rpi/requirements.txt
pydantic
rpi.gpio
pyyaml